Fix maybelle backup key fetch by using localhost delegation
Changed the task to delegate_to localhost instead of
maybelle.cryptograss.live. Since ansible is already running on maybelle,
we don't need to SSH back to it - just read the file locally.

This fixes the 'Permission denied (publickey,password)' error that was
happening when trying to SSH from hunter back to maybelle.